Extract monto_a_texto_sv into utils module
